% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{Print the top left hand corner of a matrix of data frame}
\usage{
corner(x, n = 6, n_row = NULL, n_col = NULL)
}
\arguments{
\item{x}{matrix or data.frame, object to print top corner from}

\item{n}{numeric, how many rows and columns to print}

\item{n_row}{numeric, how many rows to print, overwrites n}

\item{n_col}{numeric, how many cols to print, overwrites n}
}
\value{
top left corner of x
}
\description{
Print the top left hand corner of a matrix of data frame
}
