% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envobj.R
\name{envobj}
\alias{envobj}
\title{Print information about the objects in the current environment}
\usage{
envobj(sort = TRUE, colour = TRUE)
}
\arguments{
\item{sort}{logical, if printing results in a formatted table should it be sorted by object type (default: TRUE)}

\item{colour}{logical, should the results be printed as a colour coded table based on object type (default: TRUE)}
}
\value{
details of each object in the environment including name, type/class and dimensions will be printed to the screen
}
\description{
Print information about the objects in the current environment
}
\examples{
envobj()
}
